pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sv-agent"
version = "0.1.0"
description = "SmartSnap"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "Youtu-Agent Team"}
]
keywords = ["mobile", "automation", "android", "testing", "agent"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "backoff==2.2.1",
    "InquirerPy==0.3.4",
    "lxml>=4.9.0",
    "openai",
    "jsonlines",
    "numpy==2.1.0",
    "protobuf",
    "beautifulsoup4",
    "toml",
    "aioconsole",
    "python-dotenv",
    "gymnasium",
    "Pillow",
    "evaluate",
    "types-tqdm",
    "tiktoken",
    "aiolimiter",
    "beartype==0.12.0",
    "flask",
    "nltk",
    "text-generation",
    "opencv-python",
    "colorama",
    "dashscope",
    "pyshine",
    "pyyaml",
    "requests",
    "Levenshtein",
    "zhipuai",
    "xmltodict",
    "google-auth",
    "docker",
    "fuzzywuzzy",
    "openpyxl",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "flake8",
    "mypy",
]

[project.urls]
Homepage = "https://github.com/your-org/sv-agent"
Repository = "https://github.com/your-org/sv-agent"
Issues = "https://github.com/your-org/sv-agent/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["svagent*"]

[tool.setuptools.package-data]
"*" = ["*.txt", "*.md", "*.cfg", "*.json", "*.yaml", "*.yml"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
